Move thunk method to api file
