+++ src/pages/ClubPage.tsx
@@ -3,8 +3,8 @@ import { useDispatch, useSelector } from 'react-redux';
 import { useParams } from 'react-router-dom';
 import ClubCard from '../components/clubCard';
 import { CLUB } from '../constants';
+import { fetchDataThunk } from '../services/api';
 import { combinedStateInterface } from '../store/store';
-import { fetchDataThunk } from '../store/thunks/thunkActions';
 
 interface urlParams {
     Region: string;
