+++ src/store/thunks/thunkActions.ts
@@ -1,69 +1,3 @@
-import { Action } from 'redux';
-import { ThunkAction } from 'redux-thunk';
-import { cardType, CITY, CLUB, REGION, SPORT, TEAM } from '../../constants';
-import { fetchData } from '../../services/api';
-import { setRegionsActionCreator } from '../pages/region/regionActions';
-import { setSportsActionCreator } from '../pages/sport/sportActions';
-import { setTeamsActionCreator } from '../pages/team/teamActions';
-import { setClubsActionCreator } from '../pages/club/clubActions'
-import { combinedStateInterface } from '../store';
-import { setCitiesActionCreator } from '../pages/city/cityActions';
-
-
-export const fetchDataThunk = (
-    dataType: cardType
-): ThunkAction<void, combinedStateInterface, unknown, Action<string>> => async dispatch => {
-    dispatch(fetchInProgressActionCreator());
-    const asyncResp = await fetchData(dataType);
-     //TODO: ADD VALIDATION before dispatch
-    if (asyncResp === 'Something went wrong' || asyncResp === 'Connection error') {
-        dispatch(fetchFailedActionCreator());
-    }
-    else {
-        dispatch(fetchSuccessActionCreator());
-         //TODO: ADD VALIDATION before dispatch
-        switch (dataType) {
-            case REGION: {
-                dispatch(
-                    setRegionsActionCreator(asyncResp.results)
-                )
-                break;
-            }
-            case SPORT: {
-                dispatch(
-                    setSportsActionCreator(asyncResp.results)
-                )
-                break;
-            }
-            case CLUB: {
-                dispatch(
-                    setClubsActionCreator(asyncResp.results)
-                )
-                break;
-            }
-            case TEAM: {
-                dispatch(
-                    setTeamsActionCreator(asyncResp.results)
-                )
-                break;
-            }
-            case CITY: {
-                dispatch(
-                    setCitiesActionCreator(asyncResp.results)
-                )
-                break;
-            }
-            default: {
-                return;
-                //TODO: add error
-            }
-
-        }
-    }
-}
-
-
-
 export const FETCH_IN_PROGRESS = 'FETCH_IN_PROGRESS';
 export const FETCH_FAILED = 'FETCH_FAILED';
 export const FETCH_SUCCESS = 'FETCH_SUCCESS';
